Fixed bug in SDK build
git-svn-id: file:///afs/slac/g/reseng/svn/repos@12085 0848cbd7-7569-0410-ab05-8cc0a0c67708

Former-commit-id: 90ee8595307d81f071962d6e59907c03ace1d079
Former-commit-id: e29c7ecb151fcf92881b4e52ea29f1b8f0a135e7
Former-commit-id: f900e1f5e4d998efdddd2c6f89df864d7159d0c3

+++ StdLib/trunk/build/vivado_sdk_prj_v1.tcl
@@ -27,11 +27,11 @@ if { [file exists ${SDK_PRJ}] != 1 } {
       sdk set_workspace ${SDK_PRJ}
       sdk create_hw_project  -name hw_0  -hwspec ${SDK_PRJ}/${PROJECT}.hdf
       sdk create_bsp_project -name bsp_0 -proc microblaze_0 -hwproject hw_0 -os standalone
-      sdk create_app_project -name app_0 -app "Empty Application" -proc microblaze_0 -hwproject hw_0 -bsp bsp_0 -os standalone -lang c++
+      sdk create_app_project -name app_0 -app {Empty Application} -proc microblaze_0 -hwproject hw_0 -bsp bsp_0 -os standalone -lang c++
       file delete -force ${SDK_PRJ}/app_0/src/main.cc
       # Configure the release build
       sdk get_build_config -app app_0 build-config release
-      sdk get_build_config -app  app_0 -set compiler-optimization "Optimize for size (-Os)"      
+      sdk get_build_config -app  app_0 -set compiler-optimization {Optimize for size (-Os)}      
       foreach sdkLib ${SDK_LIB} {
          set dirName  [file tail ${sdkLib}]
          set softLink ${SDK_PRJ}/app_0/${dirName}
@@ -40,7 +40,7 @@ if { [file exists ${SDK_PRJ}] != 1 } {
       }       
       # Configure the debug build
       sdk get_build_config -app app_0 build-config debug
-      sdk get_build_config -app  app_0 -set compiler-optimization "Optimize for size (-Os)"      
+      sdk get_build_config -app  app_0 -set compiler-optimization {Optimize for size (-Os)}      
       foreach sdkLib ${SDK_LIB} {
          set dirName  [file tail ${sdkLib}]
          set softLink ${SDK_PRJ}/app_0/${dirName}
@@ -54,32 +54,33 @@ if { [file exists ${SDK_PRJ}] != 1 } {
       sdk setws ${SDK_PRJ}
       sdk createhw  -name hw_0  -hwspec ${SDK_PRJ}/${PROJECT}.hdf
       sdk createbsp -name bsp_0 -proc microblaze_0 -hwproject hw_0 -os standalone
-      sdk createapp -name app_0 -app "Empty Application" -proc microblaze_0 -hwproject hw_0 -bsp bsp_0 -os standalone -lang c++
+      sdk createapp -name app_0 -app {Empty Application} -proc microblaze_0 -hwproject hw_0 -bsp bsp_0 -os standalone -lang c++
       file delete -force ${SDK_PRJ}/app_0/src/main.cc
       # Configure the release build
       sdk configapp -app app_0 build-config release
-      sdk configapp -app  app_0 -set compiler-optimization "Optimize for size (-Os)"      
+      sdk configapp -app  app_0 -set compiler-optimization {Optimize for size (-Os)}
       foreach sdkLib ${SDK_LIB} {
          set dirName  [file tail ${sdkLib}]
          set softLink ${SDK_PRJ}/app_0/${dirName}
-         sdk configapp -app app_0 -add include-path ${sdkLib}
          exec ln -s ${sdkLib} ${softLink}
+         sdk configapp -app app_0 -add include-path ${sdkLib}
       }       
       # Configure the debug build
       sdk configapp -app app_0 build-config debug
-      sdk configapp -app  app_0 -set compiler-optimization "Optimize for size (-Os)"      
+      sdk configapp -app  app_0 -set compiler-optimization {Optimize for size (-Os)}
       foreach sdkLib ${SDK_LIB} {
-         set dirName  [file tail ${sdkLib}]
-         set softLink ${SDK_PRJ}/app_0/${dirName}
          sdk configapp -app app_0 -add include-path ${sdkLib}
-         exec ln -s ${sdkLib} ${softLink}
       }           
    }       
 
 }
 
 # Create a soft-link and add new linker to source tree
-file copy   -force ${SDK_PRJ}/app_0/src/lscript.ld ${SDK_PRJ}/app_0/lscript.ld
+if { [file exists ${SDK_PRJ}/app_0/src/lscript.ld] == 1 } {
+   exec cp -f ${SDK_PRJ}/app_0/src/lscript.ld ${SDK_PRJ}/app_0/lscript.ld
+}
 file delete -force ${SDK_PRJ}/app_0/src
 exec ln -s $::env(SDK_SRC_PATH) ${SDK_PRJ}/app_0/src
-exec mv -f ${SDK_PRJ}/app_0/lscript.ld ${SDK_PRJ}/app_0/src/lscript.ld
+if { [file exists ${SDK_PRJ}/app_0/lscript.ld] == 1 } {
+   exec mv -f ${SDK_PRJ}/app_0/lscript.ld ${SDK_PRJ}/app_0/src/lscript.ld
+}
